# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Syslogc API
  version: 0.2.0
  description: |
    REST API of Syslogc. The web UI uses only this API.

    **Authentication.** Browsers use the `slc_session` cookie set by
    `POST /api/v1/auth/login`; unsafe methods (POST, PUT, DELETE) must send the
    CSRF token from the login or `/auth/me` response in `X-CSRF-Token`.
    Automation uses API keys: `Authorization: Bearer slc_<id>_<secret>`.

    **Time ranges.** Every log read takes a `time_range` whose `from`/`to` are
    RFC 3339 timestamps or relative expressions (`now`, `now-15m`, `now-1h`,
    `now-7d`, `now/d` = start of day in `tz`). Responses echo the absolute
    `resolved_range`; send it back unchanged in parallel requests so all panels
    cover exactly the same window.

    **Errors** use RFC 9457 problem details (`application/problem+json`).
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /
security:
  - sessionCookie: []
  - apiKey: []
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Log in with username and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: {type: string, minLength: 1, maxLength: 128}
                password: {type: string, minLength: 1, maxLength: 256}
      responses:
        "200":
          description: Logged in; sets the `slc_session` cookie
          headers:
            Set-Cookie: {schema: {type: string}}
          content: {application/json: {schema: {$ref: "#/components/schemas/Session"}}}
        "401": {$ref: "#/components/responses/Problem"}
        "429": {$ref: "#/components/responses/Problem"}
...
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: Current user, permissions and CSRF token
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/Session"}}}}
        "401": {$ref: "#/components/responses/Problem"}
  /api/v1/auth/me/password:
    put:
      tags: [auth]
      summary: Change own password (revokes other sessions)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: {type: string}
                new_password: {type: string, minLength: 12, maxLength: 256}
      responses:
        "204": {description: Password changed}
        "401": {$ref: "#/components/responses/Problem"}
        "422": {$ref: "#/components/responses/Problem"}
  /api/v1/logs/search:
    post:
      tags: [logs]
      summary: Search logs (newest first, cursor pagination)
      description: |
        Requires `logs:search`; a `native` query additionally requires `logs:query_native`.
        When the native query contains pipes (`| stats …`), the response has
        `mode: table` with `columns`/`table_rows` instead of log rows.
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/SearchRequest"}}}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/SearchResponse"}}}}
        "400": {$ref: "#/components/responses/Problem"}
        "403": {$ref: "#/components/responses/Problem"}
        "422": {$ref: "#/components/responses/Problem"}
        "429": {$ref: "#/components/responses/Problem"}
        "504": {$ref: "#/components/responses/Problem"}
  /api/v1/logs/histogram:
    post:
      tags: [logs]
      summary: Log volume over time, optionally split by a field
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/HistogramRequest"}}}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/HistogramResponse"}}}}
        "422": {$ref: "#/components/responses/Problem"}
  /api/v1/logs/facets:
    post:
      tags: [logs]
      summary: Top values with exact counts for several fields
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/FacetsRequest"}}}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/FacetsResponse"}}}}
        "422": {$ref: "#/components/responses/Problem"}
  /api/v1/logs/stats:
    post:
      tags: [logs]
      summary: Aggregations (count, count_distinct, top)
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/StatsRequest"}}}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/StatsResponse"}}}}
        "422": {$ref: "#/components/responses/Problem"}
  /api/v1/logs/tail:
    get:
      tags: [logs]
      summary: Live tail (Server-Sent Events)
      description: |
        Requires `logs:tail`. Events:

        - `logs` — `data: {"rows": [LogRow, …]}` (coalesced, at most every 250 ms or 500 rows)
        - `stats` — `data: {"rows_sent": n, "dropped_client_slow": n}`
        - `error` — `data: Problem`; the stream then ends

        Comment lines (`: heartbeat`) are sent every 15 s.
      parameters:
        - name: q
          in: query
          description: base64url (no padding) encoded JSON `TailQuery`
          schema: {type: string}
        - name: start_offset
          in: query
          description: How far back to start, e.g. `5s` (default) up to `1m`.
          schema: {type: string, default: 5s}
      responses:
        "200":
          description: Event stream
          content: {text/event-stream: {schema: {type: string}}}
        "403": {$ref: "#/components/responses/Problem"}
        "429": {$ref: "#/components/responses/Problem"}
  /api/v1/logs/export:
    post:
      tags: [logs]
      summary: Streamed export (CSV, NDJSON, JSON)
      description: |
        Requires `logs:export`. The body is streamed row by row. Browsers that
        cannot stream a fetch response to disk may submit an HTML form with
        `application/x-www-form-urlencoded` fields `request` (the JSON
        `ExportRequest`) and `csrf_token`.
      parameters:
        - name: format
          in: query
          required: true
          schema: {type: string, enum: [csv, ndjson, json]}
      requestBody:
        required: true
        content:
          application/json: {schema: {$ref: "#/components/schemas/ExportRequest"}}
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                request: {type: string}
                csrf_token: {type: string}
      responses:
        "200":
          description: Attachment
          headers:
            Content-Disposition: {schema: {type: string}}
            X-Export-Limit: {schema: {type: integer}}
          content:
            text/csv: {schema: {type: string}}
            application/x-ndjson: {schema: {type: string}}
            application/json: {schema: {type: string}}
        "403": {$ref: "#/components/responses/Problem"}
        "422": {$ref: "#/components/responses/Problem"}
  /api/v1/query/validate:
    post:
      tags: [logs]
      summary: Validate a filter/native query and return the compiled native query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filter: {$ref: "#/components/schemas/FilterExpr"}
                native: {$ref: "#/components/schemas/NativeQuery"}
      responses:
        "200": {description: Result, content: {application/json: {schema: {$ref: "#/components/schemas/ValidateResponse"}}}}
  /api/v1/fields:
    post:
      tags: [fields]
      summary: Field names present in the selection
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/Selection"}}}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/FieldsResponse"}}}}
  /api/v1/fields/{field}/values:
    post:
      tags: [fields]
      summary: Top values of one field (exact counts)
      parameters:
        - {name: field, in: path, required: true, schema: {type: string, maxLength: 256}}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/FieldValuesRequest"}}}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/FieldValuesResponse"}}}}
  /api/v1/analytics/breakdown:
    post:
      tags: [analytics]
      summary: Rank the values of a field by a metric
      description: |
        `total` counts everything matching, including groups beyond `limit`,
        so `share` is honest and the returned rows need not sum to 1.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/Selection"
                - type: object
                  required: [group_by]
                  properties:
                    group_by: {type: string, example: app_name}
                    metric: {$ref: "#/components/schemas/AnalyticsMetric"}
                    limit: {type: integer, minimum: 1, maximum: 50, default: 10}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/BreakdownResponse"}}}}
        "422": {$ref: "#/components/responses/Problem"}
  /api/v1/analytics/series:
    post:
      tags: [analytics]
      summary: A metric over time, optionally one series per field value
      description: |
        Only the top `limit` groups are returned; there is no "other" series.
        `points` aligns index-for-index with `timestamps`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/Selection"
                - type: object
                  properties:
                    group_by: {type: string, example: severity}
                    metric: {$ref: "#/components/schemas/AnalyticsMetric"}
                    buckets: {type: integer, minimum: 1, maximum: 1000, default: 120}
                    limit: {type: integer, minimum: 1, maximum: 50, default: 5}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/SeriesResponse"}}}}
        "422": {$ref: "#/components/responses/Problem"}
  /api/v1/analytics/mssql:
    post:
      tags: [analytics]
      summary: Everything the SQL Server page shows, in one answer
      description: |
        Needs a source carrying the `mssql` part of the Windows template,
        which is what writes the `mssql.*` fields this reads.

        Every number is counted from SQL Server's error numbers, because the
        sentence beside a number is written in the server's own language.
        `severe_errors` counts the I/O and consistency errors (823, 824), the
        transaction log filling (9002), memory running out (701) and the
        scheduler stalling (17883, 17884). It deliberately excludes 825,
        which is a read that succeeded after a retry: nothing was lost, so it
        is counted on its own as `read_retries`.

        `sign_ins` is zero on an instance whose login auditing was left at
        failures only, which is the default on many builds — zero there means
        "not recorded", not "nobody signed in". A zero deadlock count is
        likewise not proof: SQL Server reports 1205 to the client that lost,
        and it reaches the Windows Application channel only where the message
        is marked as logged or the deadlock trace flags are on.
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/AnalysisRequest"}}}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/MSSQLResponse"}}}}
        "422": {$ref: "#/components/responses/Problem"}
  /api/v1/analytics/iis:
    post:
      tags: [analytics]
      summary: Everything the IIS page shows, in one answer
      description: |
        Needs a source carrying the `iis` part of the Windows template, which
        is what writes the `iis.*` fields this reads.

        The class counts need not add up to `requests`: a line whose
        `iis.status` could not be read as a code between 100 and 599 belongs
        to no class, and the difference is left visible rather than absorbed
        by one of them.

        There is no average or worst response time. `iis.time_taken` arrives
        as text and the store cannot sum or average a field, so "slow" is
        answered by counting the requests at or above
        `slow_threshold_millis`, and `slow_urls` ranks URLs by how many slow
        requests they had — not by how slow they were.

        Distinct counts (`clients`, `urls`, `servers`) are counted over the
        whole resolved range, because they cannot be added across buckets.
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/IISRequest"}}}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/IISResponse"}}}}
        "422": {$ref: "#/components/responses/Problem"}
  /api/v1/dashboard/overview:
    post:
      tags: [dashboard]
      summary: Overview tiles
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/DashboardRequest"}}}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/DashboardOverview"}}}}
  /api/v1/dashboard/volume:
    post:
      tags: [dashboard]
      summary: Log volume split by severity
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/DashboardRequest"}}}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/HistogramResponse"}}}}
  /api/v1/dashboard/top:
    post:
      tags: [dashboard]
      summary: Top values of a core field
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DashboardRequest"
                - type: object
                  required: [field]
                  properties:
                    field: {type: string, enum: [hostname, app_name, source_ip, facility, severity, format, source]}
                    limit: {type: integer, minimum: 1, maximum: 50, default: 10}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/FieldValuesResponse"}}}}
  /api/v1/dashboard/ingestion-rate:
    post:
      tags: [dashboard]
      summary: Ingestion rates over time from node statistics (max 24 h)
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/DashboardRequest"}}}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/IngestionRateResponse"}}}}
  /api/v1/saved-searches:
    get:
      tags: [saved-searches]
      summary: List saved searches visible to the user
      parameters:
        - {name: q, in: query, schema: {type: string}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
        - {name: cursor, in: query, schema: {type: string}}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: {type: array, items: {$ref: "#/components/schemas/SavedSearch"}}
                  next_cursor: {type: [string, "null"]}
    post:
      tags: [saved-searches]
      summary: Create a saved search
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/SavedSearchInput"}}}
      responses:
        "201": {description: Created, content: {application/json: {schema: {$ref: "#/components/schemas/SavedSearch"}}}}
        "409": {$ref: "#/components/responses/Problem"}
        "422": {$ref: "#/components/responses/Problem"}
  /api/v1/saved-searches/{id}:
    parameters:
      - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      tags: [saved-searches]
      summary: Get a saved search
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/SavedSearch"}}}}
        "404": {$ref: "#/components/responses/Problem"}
    put:
      tags: [saved-searches]
      summary: Update a saved search (optimistic concurrency via `version`)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/SavedSearchInput"
                - type: object
                  required: [version]
                  properties:
                    version: {type: integer}
      responses:
        "200": {description: OK, content: {application/json: {schema: {$ref: "#/components/schemas/SavedSearch"}}}}
        "403": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
    delete:
      tags: [saved-searches]
      summary: Delete a saved search
      responses:
        "204": {description: Deleted}
        "403": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
...
  /api/v1/api-keys/{id}:
    delete:
      tags: [api-keys]
      summary: Revoke an API key
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "204": {description: Revoked}
        "404": {$ref: "#/components/responses/Problem"}
...
  /api/v1/users/{id}/revoke-sessions:
    post:
      tags: [users]
      summary: Sign a user out of every session
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "204": {description: Sessions revoked}
...
components:
  securitySchemes:
    sessionCookie: {type: apiKey, in: cookie, name: slc_session}
    apiKey: {type: http, scheme: bearer, description: "API key `slc_<id>_<secret>`"}
  responses:
    Problem:
      description: Problem details
      content:
        application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}
  schemas:
...
    Session:
      type: object
      required: [user, permissions, csrf_token]
      properties:
        user: {$ref: "#/components/schemas/User"}
        permissions: {type: array, items: {$ref: "#/components/schemas/Permission"}}
        csrf_token: {type: string}
...
    NativeQuery:
      type: object
      required: [dialect, text]
      properties:
        dialect: {type: string, const: logsql}
        text: {type: string, maxLength: 16384}
...
    SearchRequest:
      allOf:
        - $ref: "#/components/schemas/Selection"
        - type: object
          properties:
            fields: {type: array, maxItems: 100, items: {type: string}, description: Projection (storage field names); omit for all fields}
            limit: {type: integer, minimum: 1, maximum: 10000, default: 200}
            cursor: {type: [string, "null"]}
...